Added Hashing Methods to language.java
Simple update allowing for integer-String mappings via the language
class. Might be smart to instantiate a language class for each area in
the pylon. Might be smart to make a way to map out which areas of the
pylon are being used and which are not.

Need to add method to take from a region in the pylon and load it in a
slave deflector in the partition through a language. Needs to have a
scrolling implementation.