Allocations and Memory
This week, we have a new object: an Allocation. It specifies where a
file begins and ends in a pylon. (Potentially spanning multiple levels.)
It also keeps a hashtable specific to that file.

As a result, the pylon, partition, and cursor objects have been severely
edited (for the better).

I manually imported AList and a few extensions of AList from the AList
branch, because they are just better.

Also, the Timer.java object is a simple object that can be quite useful
when it comes to runtime testing.